+++ src/RadixDlt.NetworkGateway.PostgresIntegration/LedgerExtension/ReadHelper.cs
@@ -351,7 +351,7 @@ INNER JOIN LATERAL (
         var cd = new CommandDefinition(
             commandText: @"
 SELECT
-    nextval('component_entity_state_history_id_seq') AS ComponentEntityStateHistorySequence,
+    nextval('entity_state_history_id_seq') AS EntityStateHistorySequence,
     nextval('entities_id_seq') AS EntitySequence,
     nextval('entity_access_rules_chain_history_id_seq') AS EntityAccessRulesChainHistorySequence,
     nextval('entity_metadata_history_id_seq') AS EntityMetadataHistorySequence,