+++ src/RadixDlt.NetworkGateway.PostgresIntegration/Services/EntityStateQuerier.cs
@@ -86,6 +86,8 @@ internal class EntityStateQuerier : IEntityStateQuerier
 {
     private record MetadataViewModel(long EntityId, string[] Keys, string[] Values, int TotalCount);
 
+    private record ValidatorCurrentStakeViewModel(long ValidatorId, string Balance, string State);
+
     private record FungibleViewModel(GlobalAddress ResourceEntityGlobalAddress, string Balance, int ResourcesTotalCount);
 
     private record FungibleVaultViewModel(GlobalAddress ResourceEntityGlobalAddress, string VaultAddress, string Balance, int ResourceTotalCount, int VaultTotalCount);
@@ -125,7 +127,8 @@ internal class EntityStateQuerier : IEntityStateQuerier
             (byte)_networkConfigurationProvider.GetAddressTypeDefinition(AddressSubtype.EddsaEd25519VirtualIdentityComponent).AddressBytePrefix;
     }
 
-    public async Task<GatewayModel.EntityResourcesResponse> EntityResourcesSnapshot(GlobalAddress address, bool aggregatePerVault, GatewayModel.LedgerState ledgerState, CancellationToken token = default)
+    public async Task<GatewayModel.EntityResourcesResponse> EntityResourcesSnapshot(GlobalAddress address, bool aggregatePerVault, GatewayModel.LedgerState ledgerState,
+        CancellationToken token = default)
     {
         var entity = await GetEntity<ComponentEntity>(address, ledgerState, token);
 
@@ -213,8 +216,8 @@ internal class EntityStateQuerier : IEntityStateQuerier
                 var package = await _dbContext.Entities
                     .FirstAsync(e => e.Id == ce.PackageId, token);
 
-                var state = await _dbContext.ComponentEntityStateHistory
-                    .Where(e => e.FromStateVersion <= ledgerState.StateVersion && e.ComponentEntityId == ce.Id)
+                var state = await _dbContext.EntityStateHistory
+                    .Where(e => e.FromStateVersion <= ledgerState.StateVersion && e.EntityId == ce.Id)
                     .OrderByDescending(e => e.FromStateVersion)
                     .FirstAsync(token);
 
@@ -369,7 +372,8 @@ LIMIT 1
             immutableDataHex: data.ImmutableData.ToHex());
     }
 
-    public async Task<GatewayModel.StateValidatorsListResponse> StateValidatorsList(GatewayModel.StateValidatorsListCursor? cursor, GatewayModel.LedgerState ledgerState, CancellationToken token = default)
+    public async Task<GatewayModel.StateValidatorsListResponse> StateValidatorsList(GatewayModel.StateValidatorsListCursor? cursor, GatewayModel.LedgerState ledgerState,
+        CancellationToken token = default)
     {
         var fromStateVersion = cursor?.StateVersionBoundary ?? 0;
 
@@ -399,22 +403,30 @@ LIMIT 1
         var validatorIds = validatorsAndOneMore.Take(ValidatorsLimit).Select(e => e.Id).ToArray();
 
         // TODO Validators are currently not a derived type of ComponentEntity but this is only temporary solution, remove this comment once they regain their Component-status
-        var stateById = await _dbContext.ComponentEntityStateHistory
-            .FromSqlInterpolated($@"
-WITH variables (validator_entity_id) AS (
-    SELECT UNNEST({validatorIds})
-)
-SELECT cesh.*
+        var cd = new CommandDefinition(
+            commandText: @"
+WITH variables (validator_entity_id) AS (SELECT UNNEST(@validatorIds))
+SELECT e.id as ValidatorId, CAST(evh.balance AS text), esh.state
 FROM variables
-INNER JOIN LATERAL (
-    SELECT *
-    FROM component_entity_state_history
-    WHERE component_entity_id = variables.validator_entity_id AND from_state_version <= {ledgerState.StateVersion}
-    ORDER BY from_state_version DESC
-    LIMIT 1
-) cesh ON true;
-")
-            .ToDictionaryAsync(e => e.ComponentEntityId, token);
+         INNER JOIN entities e ON e.id = variables.validator_entity_id AND from_state_version <= @stateVersion
+         INNER JOIN LATERAL ( SELECT balance
+                              FROM entity_vault_history
+                              WHERE vault_entity_id = e.stake_vault_entity_id AND from_state_version <= @stateVersion
+                              ORDER BY from_state_version DESC
+                              LIMIT 1 ) evh ON TRUE
+         INNER JOIN LATERAL ( SELECT *
+                              FROM entity_state_history
+                              WHERE entity_id = variables.validator_entity_id AND from_state_version <= @stateVersion
+                            ORDER BY from_state_version DESC
+                            LIMIT 1 ) esh ON true
+;",
+            parameters: new
+            {
+                validatorIds = validatorIds, stateVersion = ledgerState.StateVersion,
+            },
+            cancellationToken: token);
+
+        var validatorsDetails = (await _dbContext.Database.GetDbConnection().QueryAsync<ValidatorCurrentStakeViewModel>(cd)).ToList();
 
         var metadataById = await GetMetadataSlices(validatorIds, 0, DefaultMetadataLimit, ledgerState, token);
 
@@ -424,17 +436,22 @@ INNER JOIN LATERAL (
             {
                 GatewayModel.ValidatorCollectionItemActiveInEpoch? activeInEpoch = null;
 
-                if (activeSetById.TryGetValue(v.Id, out var asv))
+                if (activeSetById.TryGetValue(v.Id, out var validatorActiveSetHistory))
                 {
-                    var stake = new GatewayModel.TokenAmount(asv.Stake.ToString(), _networkConfigurationProvider.GetWellKnownAddresses().Xrd);
-                    var stakePercentage = double.Parse((asv.Stake * _tokenAmount100 / totalStake).ToString());
-
                     activeInEpoch = new GatewayModel.ValidatorCollectionItemActiveInEpoch(
-                        new GatewayModel.ValidatorCollectionItemActiveInEpochStake(stake, stakePercentage),
-                        asv.PublicKey.ToGatewayPublicKey());
+                        validatorActiveSetHistory.Stake.ToString(),
+                        double.Parse((validatorActiveSetHistory.Stake * _tokenAmount100 / totalStake).ToString()),
+                        validatorActiveSetHistory.PublicKey.ToGatewayPublicKey());
                 }
 
-                return new GatewayModel.ValidatorCollectionItem(v.GlobalAddress, new JRaw(stateById[v.Id].State), activeInEpoch, metadataById[v.Id]);
+                var details = validatorsDetails.Single(x => x.ValidatorId == v.Id);
+
+                return new GatewayModel.ValidatorCollectionItem(
+                    v.GlobalAddress,
+                    new JRaw(details.State),
+                    TokenAmount.FromSubUnitsString(details.Balance).ToString(),
+                    activeInEpoch,
+                    metadataById[v.Id]);
             })
             .ToList();
 
@@ -508,7 +525,8 @@ INNER JOIN LATERAL (
     /// <summary>
     /// Returns a paginable resource collection with total (aggregated) balance per resource.
     /// </summary>
-    private async Task<GatewayModel.FungibleResourcesCollection> GetFungiblesSlicePerResource(long entityId, int offset, int limit, GatewayModel.LedgerState ledgerState, CancellationToken token)
+    private async Task<GatewayModel.FungibleResourcesCollection> GetFungiblesSlicePerResource(long entityId, int offset, int limit, GatewayModel.LedgerState ledgerState,
+        CancellationToken token)
     {
         var cd = new CommandDefinition(
             commandText: @"
@@ -571,7 +589,8 @@ INNER JOIN entities e ON ah.fungible_resource_entity_id = e.id
     /// Returns a first page of paginable resource collection with nested paginable collection of vaults with their balances per resource for a given entity.
     /// Not suitable for "give me next page" requests.
     /// </summary>
-    private async Task<GatewayModel.FungibleResourcesCollection> GetFungiblesSlicePerVault(long entityId, int resourceLimit, int vaultLimit, GatewayModel.LedgerState ledgerState, CancellationToken token)
+    private async Task<GatewayModel.FungibleResourcesCollection> GetFungiblesSlicePerVault(long entityId, int resourceLimit, int vaultLimit, GatewayModel.LedgerState ledgerState,
+        CancellationToken token)
     {
         var cd = new CommandDefinition(
             commandText: @"
@@ -661,7 +680,8 @@ INNER JOIN entities ev ON vah.vault_entity_id = ev.id;
     /// <summary>
     /// Returns a paginable resource collection with total (aggregated) count per resource.
     /// </summary>
-    private async Task<GatewayModel.NonFungibleResourcesCollection> GetNonFungiblesSlicePerResource(long entityId, int offset, int limit, GatewayModel.LedgerState ledgerState, CancellationToken token)
+    private async Task<GatewayModel.NonFungibleResourcesCollection> GetNonFungiblesSlicePerResource(long entityId, int offset, int limit, GatewayModel.LedgerState ledgerState,
+        CancellationToken token)
     {
         var cd = new CommandDefinition(
             commandText: @"
@@ -724,7 +744,8 @@ INNER JOIN entities e ON ah.non_fungible_resource_entity_id = e.id;
     /// Returns a first page of paginable resource collection with nested paginable collection of vaults with their total count per resource for a given entity.
     /// Not suitable for "give me next page" requests.
     /// </summary>
-    private async Task<GatewayModel.NonFungibleResourcesCollection> GetNonFungiblesSlicePerVault(long entityId, int resourceLimit, int vaultLimit, GatewayModel.LedgerState ledgerState, CancellationToken token)
+    private async Task<GatewayModel.NonFungibleResourcesCollection> GetNonFungiblesSlicePerVault(long entityId, int resourceLimit, int vaultLimit,
+        GatewayModel.LedgerState ledgerState, CancellationToken token)
     {
         var cd = new CommandDefinition(
             commandText: @"
@@ -808,10 +829,12 @@ INNER JOIN entities ev ON vah.vault_entity_id = ev.id;
             ? new GatewayModel.EntityFungiblesCursor(resourceLimit).ToCursorString()
             : null;
 
-        return new GatewayModel.NonFungibleResourcesCollection(resourcesTotalCount, null, nextCursor, resources.Values.Cast<GatewayModel.NonFungibleResourcesCollectionItem>().ToList());
+        return new GatewayModel.NonFungibleResourcesCollection(resourcesTotalCount, null, nextCursor,
+            resources.Values.Cast<GatewayModel.NonFungibleResourcesCollectionItem>().ToList());
     }
 
-    private async Task<GatewayModel.NonFungibleIdsCollection> GetNonFungibleIdsSlice(long entityId, long resourceEntityId, int offset, int limit, GatewayModel.LedgerState ledgerState, CancellationToken token)
+    private async Task<GatewayModel.NonFungibleIdsCollection> GetNonFungibleIdsSlice(long entityId, long resourceEntityId, int offset, int limit,
+        GatewayModel.LedgerState ledgerState, CancellationToken token)
     {
         // TODO NG-283 this query does not account for multiple Vaults that could store NFIDs